requirements.txt
ultralytics
transformers
torch
datasets
pillow
fastapi
uvicorn[standard]